Add makeUserContext utility to create UserContext payload to sendTelemetryEvent
