+++ server/aws-lsp-codewhisperer/src/language-server/qChatServer.ts
@@ -2,7 +2,7 @@ import { CredentialsProvider, InitializeParams, Server } from '@aws/language-ser
 import { ChatController } from './chat/chatController'
 import { ChatSessionManagementService } from './chat/chatSessionManagementService'
 import { CLEAR_QUICK_ACTION, HELP_QUICK_ACTION } from './chat/quickActions'
-import { getUserAgent } from './utils'
+import { getUserAgent } from './utilities/telemetryUtils'
 
 export const QChatServer =
     (service: (credentialsProvider: CredentialsProvider) => ChatSessionManagementService): Server =>
