+++ server/aws-lsp-codewhisperer/src/language-server/utils.ts
@@ -1,13 +1,8 @@
-import {
-    InitializeParams,
-    BearerCredentials,
-    CredentialsProvider,
-} from '@aws/language-server-runtimes/server-interface'
+import { BearerCredentials, CredentialsProvider } from '@aws/language-server-runtimes/server-interface'
 import { AWSError } from 'aws-sdk'
 import { Suggestion } from './codeWhispererService'
 import { CodewhispererCompletionType } from './telemetry/types'
 import { MISSING_BEARER_TOKEN_ERROR } from './constants'
-import { ServerInfo } from '@aws/language-server-runtimes/server-interface/runtime'
 
 export function isAwsError(error: unknown): error is AWSError {
     if (error === undefined) {
@@ -86,48 +81,4 @@ export const flattenMetric = (obj: any, prefix = '') => {
     })
 
     return flattened
-}
-
-const USER_AGENT_PREFIX = 'AWS-Language-Servers'
-export const getUserAgent = (initializeParams: InitializeParams, serverInfo?: ServerInfo): string => {
-    const format = (s: string) => s.replace(/\s/g, '-')
-
-    const items: String[] = []
-
-    items.push(USER_AGENT_PREFIX)
-
-    // Fields specific to runtime artifact
-    if (serverInfo?.name) {
-        serverInfo.version
-            ? items.push(`${format(serverInfo.name)}/${serverInfo.version}`)
-            : items.push(format(serverInfo.name))
-    }
-
-    // Compute client-specific suffix
-    // Missing required data fields are replaced with 'UNKNOWN' token
-    // Whitespaces in product.name and platform.name are replaced to '-'
-    if (initializeParams?.initializationOptions?.aws) {
-        const { clientInfo } = initializeParams?.initializationOptions?.aws
-        const { extension } = clientInfo || {}
-
-        if (extension) {
-            items.push(`${extension.name ? format(extension.name) : 'UNKNOWN'}/${extension.version || 'UNKNOWN'}`)
-        }
-
-        if (clientInfo) {
-            items.push(`${clientInfo.name ? format(clientInfo.name) : 'UNKNOWN'}/${clientInfo.version || 'UNKNOWN'}`)
-        }
-
-        if (clientInfo?.clientId) {
-            items.push(`ClientId/${clientInfo?.clientId}`)
-        }
-    } else {
-        // Default to standard InitializeParams.clientInfo if no custom aws.clientInfo is set
-        const { clientInfo } = initializeParams || {}
-        if (clientInfo) {
-            items.push(`${clientInfo.name ? format(clientInfo.name) : 'UNKNOWN'}/${clientInfo.version || 'UNKNOWN'}`)
-        }
-    }
-
-    return items.join(' ')
 }
